backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
firebase-admin==6.2.0
stripe==7.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
youtube-transcript-api==0.6.1
pytube==15.0.0
whisper==1.1.10
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
redis==5.0.1
slowapi==0.1.9
langchain==0.0.350
langchain-openai==0.0.2
langchain-pinecone==0.0.1
langgraph==0.0.20
neo4j==5.15.0
graphviz==0.20.1
# Voice AI dependencies
faster-whisper==0.10.0
torch==2.1.1
torchaudio==2.1.1
pydub==0.25.1
numpy==1.24.3
# OpenVoice dependencies (placeholder - implement based on your choice)
# openvoice==0.1.0
# elevenlabs==0.2.26
# azure-cognitiveservices-speech==1.34.0
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
websockets==12.0 